Stednja memorije.
git-svn-id: svn+ssh://login.nxlab.fer.hr/fpgasvn/f32c@937 6164e1d0-f123-df11-80f4-0019d13f85ef

+++ src/lib/sio.c
@@ -8,9 +8,9 @@
 
 void (*sio_idle_fn)(void) = NULL;
 static char sio_rxbuf[SIO_RXBUFSIZE];
-static int sio_rxbuf_head = 1;
-static int sio_rxbuf_tail = 1;
-static int sio_tx_xoff = 0;
+static uint8_t sio_rxbuf_head = 1;
+static uint8_t sio_rxbuf_tail = 1;
+static uint8_t sio_tx_xoff = 0;
 
 
 static int
